Breaking Change: Types come out of /types instead of /factories
