+++ src/index.ts
@@ -7,6 +7,9 @@ export const errors = errorsImport;
 import * as validationImport from "./validation";
 export const validation = validationImport;
 
+import * as typesImport from "./types";
+export const types = typesImport;
+
 export default {
   factories,
   errors,
