+++ src/utilities/hash.ts
@@ -1,5 +1,5 @@
 import { keccak256, Message } from "js-sha3";
-import { HexString } from "./types";
+import { HexString } from "../types";
 /**
  * hash() takes a string and returns a keccak256 hash string with a 0x prefix.
  *
