--- src/utilities/types.ts
@@ -1,11 +1,0 @@
-/**
- * DSNPUserURI represents a URI targeting a user following the DSNP
- * [Identifiers](https://github.com/LibertyDSNP/spec/blob/main/pages/Identifiers.md)
- * specification.
- */
-export type DSNPUserURI = string;
-
-/**
- * Hexadecimal String
- */
-export type HexString = string;
